--- source/main_module/vehicle_fsm/vehicle_init.h
@@ -1,13 +1,0 @@
-#ifndef VEHICLE_INIT_H
-#define VEHICLE_INIT_H
-
-/**
- * @file vehicle_init.h
- * @brief Vehicle initialization implementation
- * 
- * @author Irving Wang (irvingw@purdue.edu)
- */
-
-void vehicle_init(void);
-
-#endif // VEHICLE_INIT_H
